# docs/swagger.yaml
basePath: /
definitions:
  github_com_techies_streamify_internal_models.UserResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_locked:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  github_com_techies_streamify_internal_utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_handler_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal_handler_auth.LoginResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/github_com_techies_streamify_internal_models.UserResponse'
    type: object
  internal_handler_auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  internal_handler_users.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    type: object
  internal_handler_users.UserListResponse:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/github_com_techies_streamify_internal_models.UserResponse'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@streamify.com
    name: API Support
    url: http://www.streamify.com/support
  description: Core authentication and streaming service for Streamify.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Streamify API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT access token + refresh token cookie
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_handler_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
      summary: User login
      tags:
      - Authentication
  /api/v1/auth/logout:
    post:
      description: Invalidate the current session and clear refresh token cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - Authentication
  /api/v1/auth/logout-all:
    post:
      description: Revoke all sessions and clear current refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
      summary: Logout from all devices
      tags:
      - Authentication
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Validate and rotate refresh token, issue a new access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
      summary: Refresh access token
      tags:
      - Authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_handler_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
      summary: User registration
      tags:
      - Authentication
  /api/v1/auth/verify:
    get:
      consumes:
      - application/json
      description: Verifies a user's email using a verification token and marks the
        user as verified
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
      summary: Verify user email token
      tags:
      - Authentication
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get a paginated list of users. Supports search by username, email,
        and phone number.
      parameters:
      - description: Max results per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      - description: Search by username (partial match)
        in: query
        name: username
        type: string
      - description: Search by email (partial match)
        in: query
        name: email
        type: string
      - description: Search by phone number (partial match)
        in: query
        name: phone_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_users.UserListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      description: Marks the authenticated user as deleted (soft delete). The user
        will no longer be able to log in, but their data is retained.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "500":
          description: Failed to delete user
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft-delete the current user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get a single user by their unique ID.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update the profile details of the currently authenticated user.
      parameters:
      - description: Profile update details
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/internal_handler_users.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /api/v1/users/{id}/lock:
    post:
      consumes:
      - application/json
      description: Locks a user account by setting is_locked to true
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lock user account
      tags:
      - Users
  /api/v1/users/{id}/unlock:
    post:
      consumes:
      - application/json
      description: Unlocks a user account by setting is_locked to false
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unlock user account
      tags:
      - Users
  /api/v1/users/old-soft-deleted:
    delete:
      description: Removes from the database all users who were previously soft-deleted
        and are eligible for permanent deletion (e.g., deleted for a certain period).
      produces:
      - application/json
      responses:
        "200":
          description: Old soft-deleted users permanently deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to permanently delete old soft-deleted users
          schema:
            $ref: '#/definitions/github_com_techies_streamify_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Permanently delete old soft-deleted users
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Please enter token in the format "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"